clipboard: Always clear mime type tables when updating clients clipboard
When mutter becomes the new shared clipboard owner, it can still
request for client content data, even if the remote desktop client is
not the shared clipboard owner anymore.
This can currently result in a crash, when using the VNC backend, e.g.
when mutter emits the selection-owner-changed signal, the VNC backend
asks for the mime type content, due to having a text mime type in the
new mime type list, but before getting that text, mutter asks for the
text mime type of the VNC client.

To fix this crash, always clear all mime type tables for mime types
advertised by the remote desktop client, when updating the clipboard of
the remote desktop client.

https://errors.ubuntu.com/problem/777e9e3b662530a64c542b79243c1d20ad428689