man: Only include man page content for compiled-in backends
Currently, the man page always contains sections for both the RDP- and
the VNC-backend, even if not all backends were enabled during
compilation.
This may lead to confusion to users, that respective backend setting
controls appear to be available in grdctl, although the respective
backend was not compiled in.

Fix this situation by only including the respective backend sections,
when the respective backends were enabled during compilation of
gnome-remote-desktop.