vnc: Suspend PipeWire stream on exit instead of flushing it
When destroying a PipeWire stream, gnome-remote-desktop needs to
ensure, that all dequeued buffers are queued again.
And while doing that, no new buffers should be dequeued.
In this situation, gnome-remote-desktop currently sets a helper
variable and flushes the PipeWire stream.
Flushing a PipeWire stream appears to ensure, that the PipeWire data
thread completes its current call, if there is one.

However, it has a downside: All dequeued buffers are queued again.
This also includes buffers, which are currently processed by the EGL
thread.
When trying to queue these buffers again and the flush call ended
already, gnome-remote-desktop can crash.

To fix this situation, simply set the PipeWire stream inactive.
This will result into the PipeWire data thread to be stopped, while it
first completes its current call, if there is one, but also in a state,
where all current dequeued buffers can still be queued.

https://errors.ubuntu.com/problem/f98ec0e377574042b5ac42523d730eff984500ba