clipboard-rdp: Also check for / separator when preparing uri-list
When preparing the uri-list for a file list of copied files, only the
actual selected files and folders need to be included in the uri-list,
since files in subfolders are automatically handled.
To check for the selected files, gnome-remote-desktop only checks for
the backslash ('\\') operator, but not also for the slash operator
('/'), which could lead to errors during the pasting operation, even
though the use of the slash operator in file lists is discouraged.

Fix this situation by also accounting for the slash operator, when
preparing uri-lists.

See also: https://github.com/FreeRDP/FreeRDP/issues/8350