Add base classes for hardware acceleration with Vulkan
When adding a zero-copy encode path using VAAPI, the frames to encode
need to be preprocessed. This includes doing damage detection, colour
conversion, and later also checking an additional chroma frame is
needed.
These actions will later happen within a SPIR-V shader with Vulkan.

To be able to use Vulkan, first a VkInstance needs to be created. Then
all physical devices can be retrieved. These physical devices can then
be checked to see what they are actually capable of.
Then a an actual VkDevice can be created. This VkDevice is then used for
any actual Vulkan usage, like dma-buf image creations and shader
invocations.
So, add a class, that does all these actions, i.e. handling the
VkInstance creation and physical device checks. The RDP server will
later make use of that class.

Add another class, where the actual VkDevice is handled itself too. The
VkDevice will later be created, when starting the renderer of a session.
The class for the VkDevice will later also maintain the loaded shader
modules.

When running command buffers on the GPU later, a device queue needs to
be specified. How many device queues are available and what capabilities
they have depends on the GPU and its driver.
For now, gnome-remote-desktop needs a device queue family, which
supports compute and transfer operations.
Specifying which device queues are wanted for usage must happen at the
creation of the VkDevice.
For convenience, add a GObject class for VkQueues too.

For the API version, use Vulkan 1.2. This is the maximum, that is
currently available on Intel Haswell GPUs. Later Vulkan versions make
some of the used extensions mandatory and automatically loads them.
However, since newer Vulkan versions are not available on these older
GPUs, simply load Vulkan 1.2 and load the extensions, that will be
required later manually.