vulkan: Add shader for main view creation
In order to be able to perform the colour conversion and damage
detection, add a shader for them. For performance reasons, both the
damage detection and colour conversion are handled within the same
shader. The reason is the reduction of global memory access. Accessing
memory within a shader is performance-wise the most heavy action.
When a pixel is read, it can be used for both the colour conversion and
the damage detection. This basically claps two flies with one clap.

For the colour conversion, use, for now, a similar handling to the AVC
CUDA kernel for NVENC. The colour conversion here does not entirely
match the documentation. However, it makes the whole implementation
easier for now. Overall, this shader will later be replaced by a shader,
that creates a stereo view. For now, just create the main view for
simplicity.
To the shader, three descriptor sets need to be bound, each containing
two bindings. For the second binding of the descriptor set, containing
the damage buffer, add a dummy buffer. For this main view shader, it
won't be used, but later in the stereo view shader, that will replace
this one. This dummy buffer simplifies the upcoming update mechanism in
the view creator, that will write the descriptor sets for the shader.

Vulkan uses SPIR-V for shaders, thus the shaders need to be compiled
with glslc, when building gnome-remote-desktop. So, add a build
dependency for it. Also, add a dependency for spirv-opt (part of
spirv-tools). This tool will optimize the compiled shader for better
performance.