hwaccel-nvidia: Simplify creation of NVENC sessions
Avoid large cleanup paths, when the creation of the NVENC session fails
by making the NvEncEncodeSession struct g_autoptr compatible and use
this autocleanup, when the actual session creation fails.
Also use 32-bit integers for the encode-width and encode-height, since
that is the type that is mostly used for these sizes in
gnome-remote-desktop.