rdp-audio-output-stream: Destroy the stream before clearing its pointer
Like with screencast streams, do not assume, that setting the stream
inactive will also always stop the PipeWire data thread.
Directly destroying the PipeWire stream also clears the stream
listener, so don't try to remove it manually, as clearing it before the
stream could potentially lead to unwanted side effects and trying to
clear the hook after the stream itself is an invalid operation.