egl-thread: Add ability to replace transfer tasks
The EGL thread handles transfer tasks, like uploading frame data to the
GPU or downloading dma-buf images to the host.
These tasks are all queued up into a queue, so before a task is
handled, all its previous tasks must be handled.
Some GPU drivers, like the Intel driver are extremely slow, and during
screencast sessions it is not rare, that multiple frames for one stream
are queued up, while the previous ones could and should be dropped, as
having a low frame latency is important in remote desktop sessions.

To be able to drop older operations, add the ability to acquire EGL
slots.
When submitting tasks to these slots, the pending frame operation is
dropped, while its callback is still called to be able to queue the
related PipeWire buffer again.
The overall order in the queue is still preserved across slots, so sync
operations are still possible to do.