rdp-pipewire-stream: Acquire EGL slot for transfer tasks
Since the EGL thread can now replace transfer tasks using EGL slots,
create one for each stream.
This eliminates the frame latency, that was caused due to lots of
pending tasks, when using slow drivers, like the Intel driver, where
previously it was not unusual to have up to 7 pending download tasks.
Do, however, not create an EGL slot for streams, when CUDA is
available, as currently gnome-remote-desktop still relies on CUDA-map
and -unmap operations for hardware acceleration with NVIDIA GPUs and
these tasks are not considered in this design.
In the future, these CUDA-map and CUDA-unmap operations will be replaced
by a different mechanism.