clipboard: Always call g_task_return_pointer() after reading content
Recent glib versions have an additional check in the handling of
GTasks, where it is checked, whether a GTask returned anything before
it is destroyed.
When no GTask-return call has been made, glib will print an annoying
warning about a potential bug, regardless of whether it is a bug to not
call g_task_return_*() or doing it.

For the clipboard handling in gnome-remote-desktop,
gnome-remote-desktop must be able to retrieve the clipboard content in
an async way, while also be able to answer clipboard content requests
before advertising new mime type lists.
As a result, gnome-remote-desktop cannot handle the read result in the
GTask callback, because in that case, the new mime type list was
already announced and the read request must have been already handled.
To be able to handle that situation, gnome-remote-desktop cannot use
any g_task_return_*() calls to return actual results.

To just get rid of this new glib warning, just call
g_task_return_pointer() with a NULL pointer in the GTask thread.
The overall behaviour stays the same and will continue to function,
however without that new annoying warning.