rdp: Be more strict with invalid monitor layouts
Currently, when a remote desktop client sends an invalid monitor layout,
gnome-remote-desktop either assesses it as valid or it ignores it.
This is not ideal, because if the layout is invalid, it is a bug in the
remote desktop client.
If the layout really changed, then users may perceive such situation,
where gnome-remote-desktop just ignores such layout, as server issue.
Especially, for future situations, where gnome-remote-desktop will have
to handle multi monitor configuration, an invalid layout can lead to
undefined behaviour, when attempting to sanitize the layout data, such
as gaps between monitors in the layout, or overlapping monitors.

To handle these situations better, always output warnings for invalid
monitor layouts and disconnect the client.