rdp-display-control: Disconnect clients, that violate message order
The display virtual channel extension is very simple: The server side
first sends the capabilities PDU and only then, the client is allowed to
send new monitor layouts.
If the client sends a layout, before the server sent the capabilities,
the client violated the protocol message order.
In such case, the client was also not able to constrain the layout to
the given capabilities.

To handle this situation, disconnect the faulty client.