build: Bump FreeRDP version requirement
FreeRDP 2.10.0 contains a very important fix ([0]) in the DRDYNVC
handling, which prevents data PDUs in DVCs (DYNVC_DATA_FIRST and
DYNVC_DATA) to be processed and forwarded to channel handlers, when the
respective channel has not been opened successfully.
This fix is a requirement for the [MS-RDPEA] SVC fallback handling, as
otherwise a security issue would be introduced, where a malicious client
could respond with a negative creation status for the DVC creation, but
continue sending PDUs of that not successfully opened DVC, leading to
undefined behaviour.
This undefined behaviour cannot be reliably prevented without the fix in
[0].
So bump the FreeRDP version accordingly.

[0]: https://github.com/FreeRDP/FreeRDP/commit/559dcf214a2a1203e43d57d2b7b048c5612b4f05