session: Move stream management to backends
Currently, all streams of a remote desktop session are managed in the
session class.
This works fine, as long as every session only uses one stream, but this
will change in the future for several reasons, like to support for
multi-monitors or RAIL, where multiple window streams need to be
managed.
However, the current behaviour has its downsides: When trying to invoke
multiple streams at once, then each invocation cannot be traced to which
record call it corresponds.
Creating a new stream after the previous call succeeded may work, but
makes the respective implementation too complicated and slower.

To fix this, do two things:
First, don't hook up the stream signals in the session, but in the
actual backend.
Second, pass for each stream an id to its record call.
When the stream has been created, that id can be used to look up its
related resources.