session: Use stream path instead of stream for abs pointer motion events
In theory, a stream could already be destroyed, before it can be used to
look up its stream path, and this may well be in the future.
So, instead of requiring the stream for absolute pointer motion events,
require just the stream path, since that is the data, that is in the end
needed anyway to submit that motion event.