rdp: Add stream owner class
In the future, gnome-remote-desktop may handle usual remote desktop
sessions and RAIL sessions.
For normal remote desktop sessions, virtual and monitor streams are
going to be used.
These will be managed by a layout manager, which will be implemented in
the next commits.

For RAIL sessions, mostly window streams are going to be created.
To be able to know, who issued a record call, add a stream owner class.
This class will become the base class for the layout manager and in the
future for the window manager too.