rdp: Add layout manager
Currently, gnome-remote-desktop can only handle one single stream.
In addition to that, if a stream is gone, because for example a monitor
changed its dimension (, which can commonly happen, when running
gnome-remote-desktop in a virtual machine,), then gnome-remote-desktop
is unable to restore that state.
Also, the whole session is focused around one surface.
gnome-remote-desktop currently cannot handle the loss of such surface.

To be able to solve these problems, add a layout manager. This manager
receives new monitor layouts and adapts the session to it.
It handles the stream creation, resizing, and destruction.

For this, add a state machine to adapt the session to new layouts:

First, inhibit any rendering operation and disallow any absolute mouse
pointer motion events.

Then, allocate missing RDP surfaces for each missing stream.
Since multi-monitor handling is not implemented yet, a reduction of
streams does not need to be implemented yet.
However, an implementation for stream reduction can be easily added in
the future.

After the surface creation, create streams for the surfaces, which don't
have a stream yet, or update the stream params for the surfaces, which
already have a stream.

When all streams have been created, uninhibit any rendering operation
and allow absolute mouse pointer motion events again.

When destroying RDP surfaces, it needs to be ensured, that these
surfaces don't have a corresponding GFX surface any more.
Since GFX surfaces are handled by the graphics thread, queue these
surface disposal operations onto it, when destroying the respective
surface context.

All layout related changes happen in the main thread. This includes:

1. Inhibiting current rendering operations, where the main thread waits
   for the graphics thread to finish its current operation.
2. Preparing all surface contexts, where all RDP surfaces and their
   contexts are created, in case more of them are needed.
3. Updating PipeWire stream parameters to use a different monitor size,
   in case of a stream already exists and just needs its parameters to
   be updated.
4. Creating new PipeWire streams, in case of additional streams are
   needed.
5. Finishing the layout change by updating the state machine to its
   initial state again and trigger the surface render sources, in case
   of a frame was already received.

However, there are events, that can happen during layout changes, and
that depend on the current state in general of the layout manager:

1. The client submits a pointer move or pointer click event. In that
   case, the current monitor layout is required in order to properly
   transform the monitor position.
   This event is handled in the socket thread.
2. The client sents a Suppress Output PDU, with the
   ALLOW_DISPLAY_UPDATES flag set.
   In that case, pending surface updates may be queued. In order to
   trigger their respective render GSource, the surface table needs to
   be iterated.
   Like the previous case, this event is handled in the socket thread.
3. The client can reset the graphics pipeline. After the graphics
   pipeline was reset, all surfaces need to be recreated and their
   contents need to be initialized.
   This is done by triggering their respective render GSource, and
   requires an iteration through the surface table.
   This action is triggered on the graphics thread.

Since all these three actions involve iterating over the surface table,
a mutex is required, because these actions don't happen in the main
thread.
Any changes to the surface table, however, happen in the main thread.