rdp-surface: Add APIs to allow inhibit rendering
When changing monitor layouts, and especially multi-monitor layouts in
the future, no render operations should happen while that is done.
Since each layout change on the server needs to be communicated to the
client with reset graphics PDU, when using the graphics pipeline, this
would introduce unwanted side effects during layout changes.

So, add APIs to allow the inhibition of rendering operations.
The layout manager will make use of them later.

The inhibition mechanism makes use of the surface mutex:

Currently, the surface mutex is locked by either the PW RT thread or the
EGL thread to update the new framebuffer, or it is locked by the graphics
thread, which consumes new and pending framebuffers.
Inbihiting rendering uses the very same mutex. This ensures, that
inhibiting rendering ensures, that the rendering is inhibited, when the
API call returns.

The graphics thread checks, whether rendering is inhibited, before it
attempts to render a frame.
This check happens, when the surface mutex is locked.