rdp-surface: Add API to trigger pending render operations
When a performing a protocol reset of the graphics pipeline, or when the
client tells the server side, that rendering should not be suppressed
any more, then render operations for all surfaces need to be triggered,
as there may be pending frames.
These render operations could theoretically be handled in one call on
the graphics thread, but this would later involve some nasty
synchronization issues.
It is just easier to just invoke a render source for each of them.
So, add the respective API for this.