rdp: Do not include monitor config header in session-rdp header
CUDA and WinPR declare some variables, which clash with each other, when
both header files are included, leading to unpleasant build errors.
To avoid a build failure for the API addition in the next commit, do not
include the monitor config header directly, since that header includes a
FreeRDP header, which includes a WinPR header, which will cause double
declaration errors, when being included in the same file as the CUDA
header.