rdp: Adapt to layout manager
With the layout manager now being implemented, adapt the session to it.
For now, gnome-remote-desktops behaviour won't change, but the session
won't be tied to the same stream and surface any more.

This also allows now to move the creation of the display update virtual
channel handling to the beginning of the session, as the layout manager
won't process new display updates, when the remote desktop session
hasn't started yet.
It will though pick up already submitted new layouts, when the remote
desktop session becomes ready.