rdp-layout-manager: Handle surface reduction in layout changes
The handling will be necessary for multi monitor handling later, since
monitor configurations can change in an already running session.

Since RDP surfaces are just a concept in gnome-remote-desktop, they can
be created and destroyed at any time.
However, GFX surfaces do exist in RDP and these are managed in the
graphics context, which runs on the graphics thread.
Thus, they need to be destroyed in the graphics thread.
This is already handled in the layout manager, when destroying a surface
context by queueing the disposal of the RDP surface (and with that its
respective GFX surfaces) to the graphics thread.

The surfaces that are destroyed are arbitrarily chosen. The reason for
this is that when the monitor config changes, then the client submits
the whole new configuration.
As a result, it is not known which monitor was removed or added, or
whether just the primary monitor changed (whose position is always at
(0, 0)).
So, for now, just remove arbitrary surfaces, when needed, to conform the
monitor count to the new monitor config.
In the future, a more selective approach can be added as a quality of
experience improvement.