session-rdp: Prevent DVC reinitializations on session shutdown
When a session is stopped from either gnome-shell or
gnome-remote-desktop, rather than by the remote desktop client, then it
is possible, that gnome-remote-desktop tears down a DVC, while the
socket thread realizes this, but does not expect, that the teardown was
intentional and therefore tries to reinititialize that DVC.
Normally, gnome-remote-desktop handles DVC teardowns during a session
while the channel mutex is locked.
However, during session teardown, that mutex is not used potentially
leading to a crash.

Fix this situation by simply locking the channel mutex, while tearing
down all DVCs during shutdown.

https://errors.ubuntu.com/problem/c5601c9154984352c1f4318fb10822da142cfc22