session-rdp: Prepone nw-auto instance creation
This is a preparatory step for adding connect-time network
autodetection.
The capability, that the client supports the network autodetect PDUs is
exchanged as part of the early capabilities.
During the connection sequence, there is one situation, where the server
can use the connect-time network autodetect PDUs, allowing the server
side to properly determine the round trip time or determine whether the
first frame updates should more compressed, while dropping the image
quality a bit.
For now, skip this optional phase and proceed with the connection
sequence, but already prepone the creation of the network autodetection
instance.