session-rdp: Hook up connect-time autodetection
With the addition of the connect-time network autodetect handling in the
previous commit, hook now that handling up.
FreeRDP first calls the OnConnectTimeAutoDetectBegin callback and check
its return value.
The value FREERDP_AUTODETECT_STATE_COMPLETE means that the connect-time
network autodetection phase is over and FreeRDP will continue with the
rest of the connection sequence.
The value FREERDP_AUTODETECT_STATE_REQUEST, on the other hand, means
that a request is still pending.

In the FREERDP_AUTODETECT_STATE_REQUEST state, FreeRDP will wait for the
response of the client. Upon receiving the response, the callback for
the respective PDU is called.
Then, FreeRDP calls the OnConnectTimeAutoDetectProgress callback.
These two actions happen until FREERDP_AUTODETECT_STATE_COMPLETE is
returned as the return value of the OnConnectTimeAutoDetectProgress
call.