session-rdp: Adjust colour depth handling to documentation
While FreeRDP2 only exposed the preferred colour depth of the client,
but not the supported colour depth of the client, FreeRDP3 exposes the
supported colour depth too.
This allows gnome-remote-desktop now to correctly handle situations,
where the user set the preferred colour depth to value lesser than 32,
while the graphics pipeline or codecs are supported.
Previously, gnome-remote-desktop had to guess the supported colour depth
to be 32, since the server side has to announce the used colour depth in
the Bitmap Capability Set.

So, instead of checking the preferred colour depth, check now the
supported colour depth of the client.
The colour depth value still has to be overridden, if it is not 32,
since that is the value, that is written to the Bitmap Capability Set.