rdp/nw-auto: Clear unanswered sequences on high accumulations
When an RDP client receives a RTTMeasureRequest, it MUST immediately
answer it.
Doing not so is a protocol violation.
Faulty or potentially malicious clients may, however, not answer any
RTTMeasureRequests.
As a result, gnome-remote-desktop would eventually run out of available
sequence numbers and loop indefinitely in order to try to find the next
free sequence number.

To get rid of this situation, simply clear all unanswered sequences,
when a large amount of unanswered sequences is detected.
In addition to that, also output a warning message.