settings: Add handover backend
This settings backend will be used by the handover daemon. It just
hardcodes enabled, view-only and screen-share-mode settings.
The RDP server-certificate and private key will be received from the
system daemon.

Now that there is a new settings backend, check if settings-user is
being used and only in that case connect to its
"notify::rdp-enabled-changed" signal.