ctl: Add --system option
This option is used to set the daemon-system settings. The settings
that the daemon-system uses are the RDP credentials and RDP cert/key.

It relies on pkexec to be used as gnome-remote-desktop user and
requires admin authentication to write to the daemon-system
configuration key file.

It can start and stop the systemd service too. In that case, it won't be
used pkexec because managing unit files already has a polkit action
which will require admin authentication.