ctl: Add --system option
This option is used to set daemon-system settings.
It relies on pkexec to be used as gnome-remote-desktop user and require admin
authentication.

The settings that daemon-system uses are RDP credentials and RDP
cert/key.

It can start and stop the systemd service too,
but when the keyfile settings backend is finished, it will
probably need to edit that rather than pkexec the headless
grd-ctl.