vnc: Ensure PipeWire buffers to be queued before destroying them
When the stream resolution changes or the stream is destroyed, PipeWire
will destroy all buffers for that stream.
Before destroying a buffer though, the remove_buffer callback is called.
This is very useful, because when a stream resolution change happens may
not always be predictable, and currently gnome-remote-desktop crashes,
when it processes a buffer in the EGL thread while it is destroyed.

To solve that crash, register for each PipeWire buffer a buffer context.
That buffer contains a mutex, which is locked, when the respective
buffer data needs to be processed in the EGL thread.
When the buffer data was processed, the mutex is unlocked again, if it
was locked before.
When destroying the buffer context, that same mutex is locked and
unlocked again.
If the buffer is about to be destroyed and the EGL thread still
processes it, that lock operation in the destroy function will wait for
the EGL thread, ensuring that no data, that is currently  in use, is
destroyed.

Origin: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/178

See: https://bugs.launchpad.net/ubuntu/+source/gnome-remote-desktop/+bug/2024238