rdp-pipewire-stream: Emit closed signal when source node is removed
In order to be able to implement a mechanism later to recreate monitor
streams, when they were closed due to e.g. monitor resolution changes,
a way to detect whether a stream was actually closed is needed.
Unfortunately, this is not visible from a stream consumer side.
In order to detect, when the stream has been closed, watch the PipeWire
source node id.
If the global_remove event is emitted for that node id, the stream was
closed.
Emit then the stream closed signal.