rdp-dsp: Make encoder creation optional
For audio input redirection, audio samples received from the client
need to be decoded.
The encode handling would here be superfluous.
For audio output redirection, the situation is similar: The encode
handling is required, but not the decode handling.

So, as a preparatory step for audio input redirection, make the encoder
creation optional.