rdp-telemetry: Remove protocol timeout source
The Telemetry channel exists just to retrieve the client performance
metrics after the remote desktop session was set up.
If the client for some reason hangs, then gnome-remote-desktop would
currently prematurely close the DVC.

However, such protocol timeout is not necessary in the case of the
Telemetry DVC:

Receiving the Telemetry PDU later won't impact the user experience in
any way and closing the DVC prematurely would just get rid of the
opportunity to get the performance metrics of the session startup.

So, remove the timeout source.