rdp-dsp: Add decode handling for A-law
A-law is one of the oldest audio codecs and its data size reduction is
only 50% compared to raw PCM samples.
However, it is, next to µ-law, the only actual codec supported by
Microsofts RDP client for iOS.
So, add the decode handling for A-law. It will be used by the audio
input redirection handling, which will be implemented in the next
commit.