rdp-monitor-config: Also determine layout offset
While monitor configs in RDP are built around the point (0, 0), that
means, that the primary monitor is placed at the position (0, 0), while
secondary monitors are placed relative to its position, the input- and
output-region are shifted in a way, that they only contain unsigned
values of their coordinates.
To be able to properly transform the mouse position, also determine the
offset between these two regions.

The calculation is actually very simple:
While the desktop size matches the size of the region extents of the
monitor config, the offset of the region layout matches the position of
the region extents.
So, instead of manually calculating the layout offset, simply use the
position of the region extents.