rdp-layout-manager: Fix calculation of surface position
The position of the monitor in the layout is built around (0, 0) and so,
it can be a negative value.
This value is represented in the data of the virtual monitor.

The surface position is always an unsigned value. It is shifted by the
layout offset.
The layout offset is a vector with each value being <= 0, which
represents the shift between the monitor layout and the input- and
output-region.

So, fix the calculation of the surface position accordingly.