rdp-surface: Add APIs to set and get surface mapping
Surfaces may in the future not only be mapped to an output, but rather
to a window or mapped to a scaled output.
Since for each of these actions different parameters are needed, simply
add a structure which describes for which map type it contains data and
the respective data itself.

This eases up the handling in the graphics pipeline later, since the
graphics pipeline then only has to look up the surface mapping data and
then map the surface depending on the content of the surface mapping
data, instead of guessing the type of the mapping operation.