rdp-graphics-pipeline: Use surface mapping to map surface to output
Instead of guessing the map type to be a MapSurfaceToOutput operation,
use the surface mapping data to determine the type of the mapping
operation.
In the future, this may also include operations to map a surface to a
RAIL window, or to map a surface to a scaled output or scaled RAIL
window.