rdp-layout-manager: Ensure that the correct desktop size is always used
In multi monitor scenarios, the size of one surface won't match the
desktop size any more.
In sessions, where virtual monitors are used, the layout manager knows
the desktop size, however, so it can notify the session to resize the
desktop.
So, notify the session the new desktop size, when the layout has been
updated.

In addition to that, also wait for each stream to be resized. This is
necessary for two reasons:
1. The surface size must match the expected size to not overlap other
   surfaces and to be adjacent to one other monitor.
2. When the remote desktop client sends a lot of new monitor configs at
   once, then gnome-remote-desktop still needs to be able ensure the
   correct monitor size. This is only possible by waiting for the new
   size.

For consistency, also let the session manager handle the case for the
desktop size, when a physical monitor is mirrored.
This way, the session won't have to differentiate between remote
assistance sessions and headless sessions, when resizing the desktop.